Factor out getOrgRepos into common-queries module.
